requirements.txt
django==4.2.4
whitenoise==6.5.0
psycopg2-binary==2.9.7
crispy-bootstrap5==0.7
Sphinx==7.2.4
sphinx_rtd_theme==1.3.0
django-ckeditor==6.7.0
lorem
gunicorn
django-cloudinary-storage==0.3.0
cloudinary==1.35.0
python-decouple==3.8
python-magic==0.4.27
django-storages==1.14